fix the issue of missing checks in integration tests
Because the tool deleted the checkpoints related to the SW architecture and
the Hygon architecture, but the cases in the integration test still retained
the tests for these two architectures, which resulted in missed detections
when running the integration test. This patch deleted the related tests.

Besides, comments are parsed to see if they start with a # symbol, they should
start with a // symbol. This patch changes it.

With this patch, the integration tests under all modules pass.

Change-Id: If39ea7e9b3866a987d956fc40ba20009ed3a6ddd