fix the bugs in integration tests under all modules (#3)
* fix the bugs in integration tests under all modules

This patch includes the following changes:
1. The main entry of the program is adjusted from <module>_scanner.py
under each module to main.py, so the test script cannot generate the
json file.
2. Fixed the import error in __main__.py under all modules.
3. Rename integration folder under docker module to integration_test to
keep naming consistent.

Based on the above changes, we have two ways to run these integration tests.
e.g., run the integration tests under cpp module:
1. Run "python3 -m cpp ./cpp/integration_test" under migrate-ease
   root folder with setting "PYTHONPATH=./migrate-ease/cpp".
2. Run "./test" script under ./migrate-ease/cpp/integration_test folder.

Change-Id: I9d2627441a4f34fd9b4cff8fdb7a5628006b9cb3

* fix the issue of missing checks in integration tests

Because the tool deleted the checkpoints related to the SW architecture and
the Hygon architecture, but the cases in the integration test still retained
the tests for these two architectures, which resulted in missed detections
when running the integration test. This patch deleted the related tests.

Besides, comments are parsed to see if they start with a # symbol, they should
start with a // symbol. This patch changes it.

With this patch, the integration tests under all modules pass.

Change-Id: If39ea7e9b3866a987d956fc40ba20009ed3a6ddd